pyproject.toml
[project]
name = "job-search-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "flask>=3.1.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
